corrected typos server is running properly
